+++ src/main/java/org/goafabric/catalogservice/service/controller/ChargeItemController.java
@@ -1,7 +1,7 @@
 package org.goafabric.catalogservice.service.controller;
 
 import org.goafabric.catalogservice.service.controller.vo.ChargeItem;
-import org.goafabric.catalogservice.service.logic.ChargeItemCatalogLogic;
+import org.goafabric.catalogservice.service.logic.ChargeItemLogic;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
@@ -10,19 +10,19 @@ import java.util.List;
 @RestController
 @RequestMapping(value = "chargeitems", produces = MediaType.APPLICATION_JSON_VALUE)
 public class ChargeItemController {
-    private final ChargeItemCatalogLogic chargeItemCatalogLogic;
+    private final ChargeItemLogic chargeItemLogic;
 
-    public ChargeItemController(ChargeItemCatalogLogic ChargeItemCatalogLogic) {
-        this.chargeItemCatalogLogic = ChargeItemCatalogLogic;
+    public ChargeItemController(ChargeItemLogic ChargeItemLogic) {
+        this.chargeItemLogic = ChargeItemLogic;
     }
 
     @GetMapping("/{id}")
     public ChargeItem getById(@PathVariable String id) {
-        return chargeItemCatalogLogic.getById(id);
+        return chargeItemLogic.getById(id);
     }
 
     @GetMapping("/findByDisplay")
     List<ChargeItem> findByDisplay(@RequestParam("display") String display) {
-        return chargeItemCatalogLogic.search(display);
+        return chargeItemLogic.search(display);
     }
 }