+++ src/main/java/org/goafabric/catalogservice/service/controller/InsuranceController.java
@@ -1,7 +1,7 @@
 package org.goafabric.catalogservice.service.controller;
 
 import org.goafabric.catalogservice.service.controller.vo.Insurance;
-import org.goafabric.catalogservice.service.logic.InsuranceCatalogLogic;
+import org.goafabric.catalogservice.service.logic.InsuranceLogic;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
@@ -10,19 +10,19 @@ import java.util.List;
 @RestController
 @RequestMapping(value = "insurances", produces = MediaType.APPLICATION_JSON_VALUE)
 public class InsuranceController {
-    private final InsuranceCatalogLogic insuranceCatalogLogic;
+    private final InsuranceLogic insuranceLogic;
 
-    public InsuranceController(InsuranceCatalogLogic insuranceCatalogLogic) {
-        this.insuranceCatalogLogic = insuranceCatalogLogic;
+    public InsuranceController(InsuranceLogic insuranceLogic) {
+        this.insuranceLogic = insuranceLogic;
     }
 
     @GetMapping("/{id}")
     public Insurance getById(@PathVariable String id) {
-        return insuranceCatalogLogic.getById(id);
+        return insuranceLogic.getById(id);
     }
 
     @GetMapping("/findByDisplay")
     List<Insurance> findByDisplay(@RequestParam("display") String display) {
-        return insuranceCatalogLogic.search(display);
+        return insuranceLogic.search(display);
     }
 }