+++ src/main/java/org/goafabric/catalogservice/service/logic/ConditioLogic.java
@@ -1,8 +1,8 @@
 package org.goafabric.catalogservice.service.logic;
 
-import org.goafabric.catalogservice.service.controller.vo.ChargeItem;
-import org.goafabric.catalogservice.service.repository.ChargeItemRepository;
-import org.goafabric.catalogservice.service.repository.entity.ChargeItemEo;
+import org.goafabric.catalogservice.service.controller.vo.Condition;
+import org.goafabric.catalogservice.service.repository.ConditionRepository;
+import org.goafabric.catalogservice.service.repository.entity.ConditionEo;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
 import org.springframework.stereotype.Component;
@@ -12,25 +12,25 @@ import java.util.List;
 
 @Component
 @Transactional
-public class ChargeItemCatalogLogic implements CrudLogic<ChargeItem> {
+public class ConditioLogic implements CrudLogic<Condition> {
     @Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
     interface BoMapper {
-        ChargeItem map(ChargeItemEo o);
-        ChargeItemEo map(ChargeItem o);
-        List<ChargeItem> map(List<ChargeItemEo> l);
+        Condition map(ConditionEo o);
+        ConditionEo map(Condition o);
+        List<Condition> map(List<ConditionEo> l);
     }
 
     private BoMapper mapper;
-    private ChargeItemRepository repository;
+    private ConditionRepository repository;
 
-    public ChargeItemCatalogLogic(BoMapper mapper, ChargeItemRepository repository) {
+    public ConditioLogic(BoMapper mapper, ConditionRepository repository) {
         this.mapper = mapper;
         this.repository = repository;
     }
 
     @Override
-    public void create(ChargeItem ChargeItem) {
-        repository.save(mapper.map(ChargeItem));
+    public void create(Condition condition) {
+        repository.save(mapper.map(condition));
     }
 
     @Override
@@ -44,13 +44,12 @@ public class ChargeItemCatalogLogic implements CrudLogic<ChargeItem> {
     }
 
     @Override
-    public ChargeItem getById(String id) {
+    public Condition getById(String id) {
         return mapper.map(repository.findById(id).get());
     }
 
     @Override
-    public List<ChargeItem> search(String search) {
+    public List<Condition> search(String search) {
         return mapper.map(repository.findByDisplayStartsWithIgnoreCase(search));
     }
-
 }