+++ src/test/java/org/goafabric/catalogservice/logic/ConditionCatalogLogicIT.java
@@ -1,6 +1,6 @@
 package org.goafabric.catalogservice.logic;
 
-import org.goafabric.catalogservice.service.controller.vo.Diagnosis;
+import org.goafabric.catalogservice.service.controller.vo.Condition;
 import org.goafabric.catalogservice.service.logic.CrudLogic;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,9 +9,9 @@ import org.springframework.boot.test.context.SpringBootTest;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
-public class DiagnosisCatalogLogicIT {
+public class ConditionCatalogLogicIT {
     @Autowired
-    private CrudLogic<Diagnosis> diagnosisCatalogLogic;
+    private CrudLogic<Condition> diagnosisCatalogLogic;
 
     @Test
     void search() {